Update to Baselibs 6.0.4
This update has a few changes. The main one is updates to make it inline
with ESMA Baselibs v6.0.4. The updates are:

* HDF5 1.10.6
* netCDF-C 4.7.3
* NCO 4.9.1
* nccmp 1.8.6.5
* fArgParse v0.9.2
* pFUnit v4.1.5

Beyond this, the `ChangeLog` was removed and the `CHANGELOG.md` from the
"stock" Baselibs was added.

Also, the `.gitignore` was updated to also ignore `pgfortran*` and
`nagfor*` directories.